# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Academic Workspace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/references:
    get:
      operationId: listReferences
      tags: [references]
      summary: List all references for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: References
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reference"
    post:
      operationId: createReference
      tags: [references]
      summary: Add a reference manually
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReferenceInput"
      responses:
        "201":
          description: Created reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reference"
  /projects/{projectId}/references/bulk:
    post:
      operationId: bulkAddReferences
      tags: [references]
      summary: Add multiple references at once
      description: Useful for bulk-adding auto-suggested references or batch import. Skips references with duplicate DOIs.
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkAddReferencesRequest"
      responses:
        "201":
          description: Created references
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reference"
  /projects/{projectId}/references/{referenceId}:
    delete:
      operationId: deleteReference
      tags: [references]
      summary: Delete a reference
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: number
        - name: referenceId
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/references/regenerate:
    post:
      operationId: regenerateBibliography
      tags: [references]
      summary: Regenerate bibliography for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                tier:
                  type: string
                  description: AI tier to use (e.g. "free", "standard", "premium"). Defaults to user's preferred tier.
                  example: "standard"
      responses:
        "200":
          description: Updated bibliography text
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BibliographyResult"
  /projects/{projectId}/references/validate:
    post:
      operationId: validateReferences
      tags: [references]
      summary: Validate all references against the citation format
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Validation results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferenceValidationResult"
  /projects/{projectId}/references/format:
    post:
      operationId: formatCSLBibliography
      tags: [references]
      summary: Format references as a CSL-formatted bibliography
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: number
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum: [APA, APA7, IEEE, Vancouver, Chicago, MLA, Harvard]
      responses:
        "200":
          description: Formatted bibliography text
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BibliographyResult"
  /references/fetch-metadata:
    post:
      operationId: fetchReferenceMetadata
      tags: [references]
      summary: Fetch reference metadata by DOI or ISBN
      description: Accepts a DOI or ISBN identifier and fetches metadata from CrossRef or Open Library.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FetchReferenceMetadataRequest"
      responses:
        "200":
          description: Fetched metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FetchedReferenceMetadata"
        "400":
          description: Unsupported or invalid identifier
        "404":
          description: Metadata not found
        "502":
          description: External API unavailable
  /references/search:
    get:
      operationId: searchReferences
      tags: [references]
      summary: Search academic papers from CrossRef
      description: Search CrossRef for academic papers by topic, keyword, or title. Returns structured metadata including authors, year, journal, and DOI.
      parameters:
        - name: q
          in: query
          required: true
          description: Search query (minimum 3 characters)
          schema:
            type: string
            minLength: 3
        - name: rows
          in: query
          required: false
          description: Number of results to return (max 50)
          schema:
            type: number
            default: 20
            maximum: 50
        - name: offset
          in: query
          required: false
          description: Number of results to skip for pagination
          schema:
            type: number
            default: 0
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrossRefSearchResponse"
        "400":
          description: Query too short (minimum 3 characters)
        "502":
          description: CrossRef API unavailable
...
components:
  schemas:
...
    Reference:
      type: object
      required: [id, projectId, title, validationStatus, createdAt]
      properties:
        id:
          type: number
        projectId:
          type: number
        title:
          type: string
        authors:
          type: ["string", "null"]
        year:
          type: ["number", "null"]
        journal:
          type: ["string", "null"]
        volume:
          type: ["string", "null"]
        issue:
          type: ["string", "null"]
        doi:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        validationStatus:
          type: string
          enum: [unverified, verified, invalid]
        usedInChapters:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        isSuggested:
          type: boolean
          description: Whether this reference was auto-suggested by CrossRef
        source:
          type: string
          enum: [manual, crossref, file]
          description: Source of the reference
    ReferenceInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        authors:
          type: string
        year:
          type: number
        journal:
          type: string
        volume:
          type: string
        issue:
          type: string
        doi:
          type: string
        url:
          type: string
        isSuggested:
          type: boolean
          default: false
          description: Whether this reference was auto-suggested by CrossRef search
        source:
          type: string
          enum: [manual, crossref, file]
          default: manual
          description: Source of the reference
    BulkAddReferencesRequest:
      type: object
      required: [references]
      properties:
        references:
          type: array
          items:
            $ref: "#/components/schemas/ReferenceInput"
          maxItems: 100
    BibliographyResult:
      type: object
      required: [bibliography]
      properties:
        bibliography:
          type: string
        format:
          type: string
    ValidationIssue:
      type: object
      required: [severity, message, field]
      properties:
        severity:
          type: string
          enum: [error, warning]
        message:
          type: string
        field:
          type: string
    ReferenceValidation:
      type: object
      required: [id, title, validation]
      properties:
        id:
          type: number
        title:
          type: string
        validation:
          type: object
          properties:
            valid:
              type: boolean
            issues:
              type: array
              items:
                $ref: "#/components/schemas/ValidationIssue"
    ReferenceValidationResult:
      type: object
      required: [format, totalReferences, results]
      properties:
        format:
          type: string
        totalReferences:
          type: number
        totalErrors:
          type: number
        totalWarnings:
          type: number
        results:
          type: array
          items:
            $ref: "#/components/schemas/ReferenceValidation"

    # ── Attachment ────────────────────────────────────────────────────────────
...
    FetchReferenceMetadataRequest:
      type: object
      required: [identifier]
      properties:
        identifier:
          type: string
          description: DOI (e.g. 10.1000/xyz123) or ISBN-10/ISBN-13
    FetchedReferenceMetadata:
      type: object
      required: [title, source]
      properties:
        title:
          type: string
        authors:
          type: ["string", "null"]
        year:
          type: ["number", "null"]
        journal:
          type: ["string", "null"]
        volume:
          type: ["string", "null"]
        issue:
          type: ["string", "null"]
        doi:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        publisher:
          type: ["string", "null"]
        source:
          type: string
          enum: [crossref, openlibrary, manual]
    CrossRefSearchResult:
      type: object
      properties:
        doi:
          type: ["string", "null"]
          description: Digital Object Identifier
        title:
          type: string
        authors:
          type: string
          description: Formatted author string (e.g., "John Doe, Jane Smith")
        year:
          type: ["number", "null"]
        journal:
          type: ["string", "null"]
          description: Journal or publication name
        volume:
          type: ["string", "null"]
        issue:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
          description: Publication type (e.g., journal-article, book, proceedings-paper)
        publisher:
          type: ["string", "null"]
        page:
          type: ["string", "null"]
          description: Page range (e.g., "123-145")
        abstract:
          type: ["string", "null"]
    CrossRefSearchResponse:
      type: object
      required: [results, totalResults, query]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/CrossRefSearchResult"
        totalResults:
          type: number
          description: Total number of results matching the query
        query:
          type: string
          description: The original search query
...